CPU가 DB Response 처리할 때 튀는 현상이 있음 해결 방안 모색 필요

+++ ServerCore/Monitor.cpp
@@ -36,14 +36,14 @@ const uint64 Monitor::GetMemoryUsage_Byte()
 	return 0;
 }
 
-const uint64 Monitor::GetMemoryUsage_KB()
+const double Monitor::GetMemoryUsage_KB()
 {
-	return GetMemoryUsage_Byte() / 1024;
+	return static_cast<double>(GetMemoryUsage_Byte()) / 1024.0;
 }
 
-const uint64 Monitor::GetMemoryUsage_MB()
+const double Monitor::GetMemoryUsage_MB()
 {
-	return GetMemoryUsage_KB() / 1024;
+	return GetMemoryUsage_KB() / 1024.0;
 }
 
 void Monitor::TimeMonitorCheck(const std::string& fn, uint64 tick)
